move modal state controls to a shared mixin
